Enable only printing directory name when we first enter a directory.
This was weird. I implemented it with a statefile because when I used
an environment variable it did not seem to consistently update as
I reset its contents. I believe this is because the functions that
populate the prompt are executing in subshells.

Reading content from a file maintains no state, though,